Object sizes should be calculated as UDATA instead of U_32

+++ debugtools/DDR_VM/src/com/ibm/j9ddr/vm29/tools/ddrinteractive/commands/ObjectSizeInfo.java
@@ -313,7 +313,7 @@ public class ObjectSizeInfo extends Command
 			
 			long accum = 0;
 			if (!superclass.isNull()) {
-				accum = superclass.totalInstanceSize().intValue();
+				accum = superclass.totalInstanceSize().longValue();
 				if (superclass.backfillOffset().gt(0)) {
 					accum -= ObjectFieldInfo.BACKFILL_SIZE;
 				}
